Fix mlnx_tune out of range tx number
In order to use set_xps_mask in the get_rings function, there must
be an existing tx queue with the same number as the rx_ring object we
use the set_xps_mask on.
This commit fix the get_rings function.
The commit add a restriction to handle the out of range problem.

Signed-off-by: Shmuel Shaul <sshaul@nvidia.com>