Change mlnx_tune get_os()
Currently, mlnx_tune hold OS name constants inside the OS class.
inside the get_os() function, it filtering the os information by using
the linux_distribution() output.
if the os name/version is not part of the constant inside the OS class
the get_os() return UNKNOWN_OS.
this commit replace the extra filtering inside get_os()
and in case of an OS that is not defined in the OS class,
it returns the os distribution and version.

Signed-off-by: Shmuel Shaul <sshaul@nvidia.com>