mlnx_qos: Do not skip first buffer size line
In previous change, mlnx_qos query was updated to print total_size as
well, which is deduced from qos sysfs output opening line. However the
readline() method moves the pointer to new line, so there is no need to
skip 3 lines as before to reach buffer size info.

Change the navigation of qos sysfs outtputs to look for lines that begin
with numbers as an indication of buffer size info data.

Signed-off-by: Maher Sanalla <msanalla@nvidia.com>