Avoid the "python(abi)" rpm generated requirement
Make the rpm package more portable for different python versions in the
same distribution:

Currently there is one script, mlnx_qos, that uses python modules from a
python-version-specific directory. The 'python(abi)' dependency is
created because of that.

Avoid creating it, as the expense of potentially breaking mlnx_qos on
such systems, but at least being able to install mlnx-tools.

Issue: 3431430
Signed-off-by: Tzafrir Cohen <nvidia@cohens.org.il>