mlnx_bf_configure: Increase NetworkManager timeout
In some cases NetworkManager is not active in 60 seconds which prevents
SFs and OVS bridges from being created. So, the timeout increased to 2
minutes and now it is configurable.

Add "NETWORK_MANAGER_TIMEOUT=<sec>" to the /etc/mellanox/mlnx-bf.conf
Also, in case of timeout continue to SF creation instead of exiting.

Issue: 3873073
Signed-off-by: Vladimir Sokolovsky <vlad@nvidia.com>