mlnx_tune: Fix SyntaxWarning: invalid escape sequence with Python3.12
Python 3.12 upgraded the DeprecationWarning to a SyntaxWarning. Unlike
DeprecationWarnings, SyntaxWarnings are displayed by default. this is
why we see these erros only now.
currently, its not damaging the functionality, but In a future version
of python, using invalid escape sequence in string literals is planned
to eventually become a hard SyntaxError, that will break the script.
Fix it by change regex strings to raw strings.

Signed-off-by: Shmuel Shaul <sshaul@nvidia.com>