tsbin/mlnx_bf_configure: Stop OVS service before changing eswitch mode
OVS creates and manages internal bridges and ports which can interfere
with changing the eswitch mode. Therefore, we need to backup the OVS
configurations, remove OVS bridges and stop the OVS service before
changing the eswitch mode. If the service was active prior to the mode
change we start it again after changing the eswitch and re-apply the old
configs.

Issue: 4409382
Signed-off-by: Stav Aviram <saviram@nvidia.com>